Boot/genVDisk/Kernel: Made some changes to fix magic string.
We had to change Bochs from using a 1.44MB floppy to a 2.88MB one, because
that is the default on QEMU and it's harder to change there. This meant
that we also needed to change the geometry defined in the bootloader. We
should really just ditch floppies completely and switch to a hard disk.